requirements.txt
cycler==0.10.0
glob2==0.6
kiwisolver==1.0.1
matplotlib==3.0.2
nibabel==2.3.1
nilearn==0.5.0
nose==1.3.7
numpy==1.15.4
opencv-contrib-python==3.4.4.19
pandas==0.23.4
pathlib==1.0.1
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
